Rename nc_nwcsaf module to nwcsaf_nc to better match reader names
