+++ satpy/readers/nwcsaf_nc.py
